style: refactor folder-centric & PBR-like
