+++ src/algebra/prelude.rs
@@ -1,3 +1,5 @@
-pub use super::bounds::BoundingBox;
 /// This is the prelude file for the Algebra core
+pub use super::bounds::BoundingBox;
+pub use super::intersection::Intersection;
+pub use super::ray::Ray;
 pub use super::vectors::{Vec2, Vec3, Vertex};
