+++ src/core.rs
@@ -1,0 +1,1 @@
+pub mod core::shape;
